@@ -1,8 +1,8 @@
-var express = require('express');
-const dbConnect = require('../routes/repository');
+import express, { Request, Response, Express, NextFunction } from 'express';
+import dbConnect from './repository';
 
-var router = express.Router();
-router.get('/cats/:id', async (req, res) => {
+const router = express.Router();
+router.get('/cats/:id', async (req: Request, res: Response) => {
   const { id } = req.params;
   console.log("-check request- \n id : ", id);
 
@@ -16,7 +16,7 @@ router.get('/cats/:id', async (req, res) => {
 
   } catch (err) {
     console.error(err);
-    res.status(500).json({ error: err.message });
+    res.status(500).json({ error: err });
     return;
 
   } finally {
@@ -30,7 +30,7 @@ router.get('/cats/:id', async (req, res) => {
   }
 })
 
-router.get('/cats', async (req, res) => {
+router.get('/cats', async (req: Request, res: Response) => {
 
   let connection;
 
@@ -42,14 +42,14 @@ router.get('/cats', async (req, res) => {
 
   } catch (err) {
     console.error(err);
-    res.status(500).json({ error: err.message });
+    res.status(500).json({ error: err });
     return;
 
   } finally {
   }
 });
 
-router.post('/cats', async (req, res) => {
+router.post('/cats', async (req: Request, res: Response) => {
   const { name, age, breed } = req.body;
   console.log("Received data:", { name, age, breed });
 
@@ -61,11 +61,11 @@ router.post('/cats', async (req, res) => {
       'INSERT INTO `cats` (name, age, breed) VALUES (?, ?, ?)',
       [name, age, breed]
     );
-    res.json({ message: 'success', data: { id: result.insertId, name, age, breed } });
+    res.json({ message: 'success', data: { id: (result as any).INSERT, name, age, breed } });
 
   } catch (err) {
     console.error(err);
-    res.status(500).json({ error: err.message });
+    res.status(500).json({ error: err });
 
   } finally {
     if (connection) {
@@ -79,7 +79,7 @@ router.post('/cats', async (req, res) => {
 
 });
 
-router.put('/cats/:id', async (res, req) => {
+router.put('/cats/:id', async (req: Request, res: Response) => {
   const { id } = req.params;
 
   let connection;
@@ -96,7 +96,7 @@ router.put('/cats/:id', async (res, req) => {
     console.log(result);
 
 
-  } catch (error) {
+  } catch (err) {
     console.error("500err", err);
     res.status(500).json({ error: "500err" });
     return;
@@ -112,7 +112,7 @@ router.put('/cats/:id', async (res, req) => {
   }
 })
 
-router.delete('/cats/:id', async (req, res) => {
+router.delete('/cats/:id', async (req: Request, res: Response) => {
   const { id } = req.params;
 
   let connection;
@@ -146,4 +146,4 @@ router.delete('/cats/:id', async (req, res) => {
   }
 });
 
-module.exports = router;
+export default router;
